+++ unnamed_engine/src/renderer/middleware_renderer.rs
@@ -3,11 +3,11 @@
 //! Defines a middleware that stores and executes everything related to the graphics library.
 use cgmath::{InnerSpace, Rotation3, Zero};
 use egui_wgpu::ScreenDescriptor;
-use wgpu::{hal::auxil::db, util::DeviceExt, BindGroup, TextureFormat};
+use wgpu::{util::DeviceExt, BindGroup};
 
 use crate::gui::{egui_renderer::EguiRenderer, gui::gui};
 
-use super::{camera::CameraController, texture::{self, Texture}, viewport::Viewport};
+use super::{camera::CameraController, texture::{self, Texture}, transform::{Transform, TransformRaw}, viewport::Viewport};
 
 #[repr(C)]
 #[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
@@ -55,63 +55,6 @@ const INDICES: &[u16] = &[
 const NUM_INSTANCES_PER_ROW: u32 = 10;
 const INSTANCE_DISPLACEMENT: cgmath::Vector3<f32> = cgmath::Vector3::new(NUM_INSTANCES_PER_ROW as f32 * 0.5, 0.0, NUM_INSTANCES_PER_ROW as f32 * 0.5);
 
-struct Instance {
-  position: cgmath::Vector3<f32>,
-  rotation: cgmath::Quaternion<f32>,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct InstanceRaw {
-  model: [[f32; 4]; 4],
-}
-
-impl Instance {
-  fn to_raw(&self) -> InstanceRaw {
-    InstanceRaw {
-      model: (cgmath::Matrix4::from_translation(self.position) * cgmath::Matrix4::from(self.rotation)).into(),
-    }
-  }
-}
-
-impl InstanceRaw {
-  fn desc() -> wgpu::VertexBufferLayout<'static> {
-      use std::mem;
-      wgpu::VertexBufferLayout {
-          array_stride: mem::size_of::<InstanceRaw>() as wgpu::BufferAddress,
-          // We need to switch from using a step mode of Vertex to Instance
-          // This means that our shaders will only change to use the next
-          // instance when the shader starts processing a new instance
-          step_mode: wgpu::VertexStepMode::Instance,
-          attributes: &[
-              // A mat4 takes up 4 vertex slots as it is technically 4 vec4s. We need to define a slot
-              // for each vec4. We'll have to reassemble the mat4 in the shader.
-              wgpu::VertexAttribute {
-                  offset: 0,
-                  // While our vertex shader only uses locations 0, and 1 now, in later tutorials, we'll
-                  // be using 2, 3, and 4, for Vertex. We'll start at slot 5, not conflict with them later
-                  shader_location: 5,
-                  format: wgpu::VertexFormat::Float32x4,
-              },
-              wgpu::VertexAttribute {
-                  offset: mem::size_of::<[f32; 4]>() as wgpu::BufferAddress,
-                  shader_location: 6,
-                  format: wgpu::VertexFormat::Float32x4,
-              },
-              wgpu::VertexAttribute {
-                  offset: mem::size_of::<[f32; 8]>() as wgpu::BufferAddress,
-                  shader_location: 7,
-                  format: wgpu::VertexFormat::Float32x4,
-              },
-              wgpu::VertexAttribute {
-                  offset: mem::size_of::<[f32; 12]>() as wgpu::BufferAddress,
-                  shader_location: 8,
-                  format: wgpu::VertexFormat::Float32x4,
-              },
-          ],
-      }
-  }
-}
 
 pub struct MiddlewareRenderer {
   depth_texture: Texture,
@@ -121,7 +64,7 @@ pub struct MiddlewareRenderer {
   pipeline: wgpu::RenderPipeline,
   vertex_buffer: wgpu::Buffer,
   index_buffer: wgpu::Buffer,
-  instances: Vec<Instance>,
+  transforms: Vec<Transform>,
   instance_buffer: wgpu::Buffer,
   num_indices: u32,
 }
@@ -196,7 +139,7 @@ impl MiddlewareRenderer {
       vertex: wgpu::VertexState {
         module: &shader,
         entry_point: "vs_main",
-        buffers: &[Vertex::desc(), InstanceRaw::desc()],
+        buffers: &[Vertex::desc(), TransformRaw::desc()],
       },
       fragment: Some(wgpu::FragmentState {
         module: &shader,
@@ -245,7 +188,7 @@ impl MiddlewareRenderer {
 
     let num_indices = INDICES.len() as u32;
 
-    let instances = (0..NUM_INSTANCES_PER_ROW).flat_map(|z| {
+    let transforms = (0..NUM_INSTANCES_PER_ROW).flat_map(|z| {
       (0..NUM_INSTANCES_PER_ROW).map(move |x| {
         let position = cgmath::Vector3 {x: x as f32, y: 0.0, z: z as f32} - INSTANCE_DISPLACEMENT;
 
@@ -255,13 +198,15 @@ impl MiddlewareRenderer {
           cgmath::Quaternion::from_axis_angle(position.normalize(), cgmath::Deg(45.0))
         };
 
-        Instance {
-          position, rotation
+        let scale  = cgmath::Vector3{x: 1.0, y: 1.0, z: 1.0};
+
+        Transform {
+          position, rotation, scale
         }
       })
     }).collect::<Vec<_>>();
 
-    let instance_data = instances.iter().map(Instance::to_raw).collect::<Vec<_>>();
+    let instance_data = transforms.iter().map(Transform::to_raw).collect::<Vec<_>>();
     let instance_buffer = device.create_buffer_init(
       &wgpu::util::BufferInitDescriptor {
         label: Some("instance_buffer"),
@@ -279,7 +224,7 @@ impl MiddlewareRenderer {
       vertex_buffer,
       index_buffer,
       num_indices,
-      instances,
+      transforms,
       instance_buffer,
     }
   }
@@ -340,7 +285,7 @@ impl MiddlewareRenderer {
       render_pass.set_vertex_buffer(1, self.instance_buffer.slice(..));
       render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
 
-      render_pass.draw_indexed(0..self.num_indices, 0, 0..self.instances.len() as _);
+      render_pass.draw_indexed(0..self.num_indices, 0, 0..self.transforms.len() as _);
     }
 
     let screen_descriptor = ScreenDescriptor {