+++ unnamed_engine/src/renderer/mod.rs
@@ -2,6 +2,7 @@
 //!
 //! The UnnamedEngine renderer.
 pub mod viewport;
+pub mod transform;
 pub mod middleware_renderer;
 pub mod texture;
 pub mod camera;